+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## [Unreleased]
 ### Added
 
-- app icon
+- App icon
+- Implemented adding Favorites
 
 ### Changed
 
@@ -13,6 +14,7 @@
 ____
 ## [0.0.10] - 2022-01-08
 ### Added
+
 - Changelog file
 - Implemented importing Annotations
 - Implemented importing Highlights
@@ -31,6 +33,7 @@ ____
 
 ## [0.0.9] - 2022-01-02
 ### Added
+
 - Implemented importing Notes
 - Implemented creating a new/blank archive
 - Added context-aware buttons to interface
@@ -39,19 +42,23 @@ ____
   - Import - for Notes, Highlights, Annotations - process for the last two not implemente yet
 
 ### Changed
+
 - Various cosmetic modifications and small bug fixes
 
 ## [0.0.8] - 2021-12-28
 ### Added
+
 - Added option to group by publication type
 
 ### Changed
+
 - Updated publications list
 - Moved resources to SQLite db
 - Improved About and Help dialogs
 
 ## [0.0.7] - 2021-12-24
 ### Added
+
 - Implemented viewing, deleting, exporting, reindexing, saving
 ____
 [Unreleased]: https://gitlab.com/erykj/jwlmanager